Move the snake case functions into their own module
